Handle unicode encoding in logged response
Previously, if a non ascii byte were present in the response for a
client query, a format string call would try to encode the response
string a second time, resulting in a unicode conversion error. In this
commit, the format log string is declared as a unicode string. This
allows the second encoding of the response string to be skipped over,
perserving the already encoded bytes and avoiding the error.